gonan try to fix the names

+++ HighwayHurry/Utils.h
@@ -17,7 +17,7 @@ void applyGlobalScale(sf::Sprite& sprite);
 /// <summary>
 /// https://stackoverflow.com/questions/14505571/centering-text-on-the-screen-with-sfml
 /// </summary>
-void centerText(sf::Text& text, const int x, const int y);
+void centerText(sf::Text& text, const float x, const float y);
 
 /// <summary>
 /// https://stackoverflow.com/questions/14505571/centering-text-on-the-screen-with-sfml
